requirements.txt
langchain==1.2.10
langchain-groq==1.1.2
pytesseract==0.3.13
Pillow==12.1.1
opencv-python-headless==4.8.1.78
python-dotenv==1.2.1
SQLAlchemy==2.0.47
numpy==2.4.2